+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * Support for UUID field types and UUID values
 
+### Fixed
+
+* Fixed select crash when dropping indexes.
+
 ## [0.4.0] - 2020-12-02
 
 ### Fixed